Get rid of test_client.py
